cmd/snap-update-ns: move user profile code to user.go
