+++ cmd/snap-update-ns/main.go
@@ -22,11 +22,9 @@ package main
 import (
 	"fmt"
 	"os"
-	"strings"
 
 	"github.com/jessevdk/go-flags"
 
-	"github.com/snapcore/snapd/dirs"
 	"github.com/snapcore/snapd/logger"
 	"github.com/snapcore/snapd/osutil"
 )
@@ -85,36 +83,6 @@ func run() error {
 		return applyUserFstab(opts.Positionals.SnapName)
 	}
 	return applySystemFstab(opts.Positionals.SnapName, opts.FromSnapConfine)
-}
-
-func applyUserFstab(snapName string) error {
-	desiredProfilePath := fmt.Sprintf("%s/snap.%s.user-fstab", dirs.SnapMountPolicyDir, snapName)
-	desired, err := osutil.LoadMountProfile(desiredProfilePath)
-	if err != nil {
-		return fmt.Errorf("cannot load desired user mount profile of snap %q: %s", snapName, err)
-	}
-
-	// Replace XDG_RUNTIME_DIR in mount profile
-	xdgRuntimeDir := fmt.Sprintf("%s/%d", dirs.XdgRuntimeDirBase, os.Getuid())
-	for i := range desired.Entries {
-		if strings.HasPrefix(desired.Entries[i].Name, "$XDG_RUNTIME_DIR/") {
-			desired.Entries[i].Name = strings.Replace(desired.Entries[i].Name, "$XDG_RUNTIME_DIR", xdgRuntimeDir, 1)
-		}
-		if strings.HasPrefix(desired.Entries[i].Dir, "$XDG_RUNTIME_DIR/") {
-			desired.Entries[i].Dir = strings.Replace(desired.Entries[i].Dir, "$XDG_RUNTIME_DIR", xdgRuntimeDir, 1)
-		}
-	}
-
-	debugShowProfile(desired, "desired mount profile")
-
-	// TODO: configure the secure helper and inform it about directories that
-	// can be created without trespassing.
-	as := &Assumptions{}
-	// TODO: Handle /home/*/snap/* when we do per-user mount namespaces and
-	// allow defining layout items that refer to SNAP_USER_DATA and
-	// SNAP_USER_COMMON.
-	_, err = applyProfile(snapName, &osutil.MountProfile{}, desired, as)
-	return err
 }
 
 func applyProfile(snapName string, currentBefore, desired *osutil.MountProfile, as *Assumptions) (*osutil.MountProfile, error) {
